test/test_view/test_tradetab.py
import unittest
from tkinter import Tk
from unittest.mock import Mock

from view.tradetab import TradeTab


class TestTradeTab(unittest.TestCase):
    def setUp(self):
        self.root = Tk()
        self.presenter = Mock()
        self.trade_tab = TradeTab(self.root, self.presenter)

    def test_init(self):
        # Ensure that the necessary attributes and widgets are initialized correctly
        self.assertIsNotNone(self.trade_tab.type_var)
        self.assertIsNotNone(self.trade_tab.exchange_var)
        self.assertIsNotNone(self.trade_tab.stoploss_slider)
        self.assertIsNotNone(self.trade_tab.takeprofit_slider)
        self.assertIsNotNone(self.trade_tab.trade_button)
        self.assertIsNotNone(self.trade_tab.stop_loss_button)
        self.assertIsNotNone(self.trade_tab.take_profit_button)
        self.assertIsNotNone(self.trade_tab.amount_label)
        self.assertIsNotNone(self.trade_tab.amount_entry)
        self.assertIsNotNone(self.trade_tab.price_label)
        self.assertIsNotNone(self.trade_tab.price_entry)
        self.assertIsNotNone(self.trade_tab.usdt_label)
        self.assertIsNotNone(self.trade_tab.btc_label)
        self.assertIsNotNone(self.trade_tab.usdt_balance_label)
        self.assertIsNotNone(self.trade_tab.btc_balance_label)

    def test_create_ui(self):
        # Test the creation of the user interface
        self.trade_tab.create_ui(self.presenter)

        # Replace the following line with your actual assertions
        self.assertTrue(True)

    # Add more test cases for other methods as needed


if __name__ == '__main__':
    unittest.main()
